Rename ableton styles
