+++ src/app/levelone/(pages)/ableton/_components_/AbletonHeader.tsx
@@ -10,7 +10,7 @@ import {
 	FaCaretDown,
 } from 'react-icons/fa';
 
-import { outfit } from './styles';
+import { outfit } from './AbletonStyles';
 import interactiveLook from './interactive.module.scss';
 import globalLook from '@/styles/globals.module.scss';
 import { SafeExternalLink } from '@/components/global/Commons';
